refactor admin area

+++ src/Clients/VoteSystem.Clients.MVC/Areas/Administration/Controllers/VoteSystemController.cs
@@ -2,8 +2,11 @@
 using System.Linq;
 using System.Web.Mvc;
 using System.Web.Mvc.Expressions;
+
 using VoteSystem.Clients.MVC.Areas.Administration.ViewModels.VoteSystem;
+using VoteSystem.Clients.MVC.Infrastructure.Attributes;
 using VoteSystem.Clients.MVC.Infrastructure.Mapping;
+using VoteSystem.Clients.MVC.Infrastructure.NotificationSystem;
 using VoteSystem.Data.Services.Contracts;
 
 namespace VoteSystem.Clients.MVC.Areas.Administration.Controllers
@@ -20,7 +23,7 @@ namespace VoteSystem.Clients.MVC.Areas.Administration.Controllers
         }
 
         [HttpGet]
-        public ActionResult Index()
+        public ViewResult Index()
         {
             var allVoteSystems = _voteSystemService
                                                 .All()
@@ -31,7 +34,7 @@ namespace VoteSystem.Clients.MVC.Areas.Administration.Controllers
         }
 
         [HttpGet]
-        public ActionResult Create()
+        public ViewResult Create()
         {
             return View();
         }
@@ -45,8 +48,18 @@ namespace VoteSystem.Clients.MVC.Areas.Administration.Controllers
                 return View(model);
             }
 
-            var voteSystemAsDbObject = Mapper.Map<Data.Entities.VoteSystem>(model);
-            _voteSystemService.Add(voteSystemAsDbObject);
+            try
+            {
+                var voteSystemAsDbObject = Mapper.Map<Data.Entities.VoteSystem>(model);
+                _voteSystemService.Add(voteSystemAsDbObject);
+
+                this.AddNotification("Успешно създадохте система за гласуване!", NotificationType.Success);
+            }
+            catch (Exception ex)
+            {
+                // TODO add logic logic
+                this.AddNotification("Възникна грешка при създаването на системата!", NotificationType.Error);
+            }
 
             return this.RedirectToAction<VoteSystemController>(c => c.Index());
         }
@@ -54,8 +67,13 @@ namespace VoteSystem.Clients.MVC.Areas.Administration.Controllers
         [HttpGet]
         public ActionResult Edit(int voteSystemId)
         {
-            var voteSystem = _voteSystemService.GetById(voteSystemId);
+            if (voteSystemId <= 0)
+            {
+                // TODO redirect to page 404
+                return Content("voteSystemId cannot be negative number or 0");
+            }
 
+            var voteSystem = _voteSystemService.GetById(voteSystemId);
             var voteSystemAsViewModel = Mapper.Map<VoteSystemViewModel>(voteSystem);
 
             return View(voteSystemAsViewModel);
@@ -70,9 +88,17 @@ namespace VoteSystem.Clients.MVC.Areas.Administration.Controllers
                 return View(model);
             }
 
-            var voteSystemAsDbObject = Mapper.Map<Data.Entities.VoteSystem>(model);
+            try
+            {
+                var voteSystemAsDbObject = Mapper.Map<Data.Entities.VoteSystem>(model);
+                _voteSystemService.Update(voteSystemAsDbObject);
 
-            _voteSystemService.Update(voteSystemAsDbObject);
+                this.AddNotification("Успешно редактирахте системата за гласуване!", NotificationType.Success);
+            }
+            catch (Exception ex)
+            {
+                this.AddNotification("Възникна грешка при редактирането на системата", NotificationType.Error);
+            }
 
             return this.RedirectToAction<VoteSystemController>(c => c.Index());
         }
@@ -80,6 +106,12 @@ namespace VoteSystem.Clients.MVC.Areas.Administration.Controllers
         [HttpPost]
         public ContentResult Delete(int voteSystemId)
         {
+            if (voteSystemId <= 0)
+            {
+                // TODO redirect to page 404
+                return Content("voteSystemId cannot be negative number or 0");
+            }
+
             _voteSystemService.Delete(voteSystemId);
 
             return Content("DELETED");
@@ -88,15 +120,22 @@ namespace VoteSystem.Clients.MVC.Areas.Administration.Controllers
         [HttpGet]
         public ActionResult Result(int voteSystemId)
         {
+            if (voteSystemId <= 0)
+            {
+                // TODO redirect to page 404
+                return Content("voteSystemId cannot be negative number or 0");
+            }
+
             return View(voteSystemId);
         }
 
         [HttpGet]
+        [AjaxOnly]
         public JsonResult GetVoteSystemResultsById(int voteSystemId)
         {
             if (voteSystemId <= 0)
             {
-                return Json("voteSystemId can not be negative number or 0", JsonRequestBehavior.AllowGet);
+                return Json("-1", JsonRequestBehavior.AllowGet);
             }
 
             var result = _questionService
